# docs/api/openapi.yaml
openapi: 3.1.0
info:
  title: TrackPick API
  version: 1.0.0
  description: Contract for the TrackPick MVP HTTP API.
  license:
    name: MIT
    identifier: MIT
servers:
  - url: /api/v1
paths:
...
  /auth/sign-up:
    post:
      operationId: signUp
      summary: Register an email account with Supabase Auth
      tags: [Identity]
      security: []
      parameters:
        - $ref: '#/components/parameters/CsrfHeader'
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/SignUpRequest'
      responses:
        '202':
          description: Registration was accepted and a verification email was requested.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/SignUpResponse'
        '400':
          $ref: '#/components/responses/ValidationError'
        '409':
          $ref: '#/components/responses/ConflictError'
        '429':
          $ref: '#/components/responses/RateLimitError'
  /auth/login:
    post:
      operationId: login
      summary: Create an authenticated session
      tags: [Identity]
      security: []
      parameters:
        - $ref: '#/components/parameters/CsrfHeader'
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/LoginRequest'
      responses:
        '200':
          description: Login succeeded.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/AccountResponse'
        '401':
          $ref: '#/components/responses/AuthenticationError'
        '403':
          $ref: '#/components/responses/ForbiddenError'
        '429':
          $ref: '#/components/responses/RateLimitError'
  /auth/password-recovery:
    post:
      operationId: requestPasswordRecovery
      summary: Send a Supabase password recovery email
      tags: [Identity]
      security: []
      parameters:
        - $ref: '#/components/parameters/CsrfHeader'
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/PasswordRecoveryRequest'
      responses:
        '202':
          description: The recovery request was accepted.
        '400':
          $ref: '#/components/responses/ValidationError'
        '429':
          $ref: '#/components/responses/RateLimitError'
  /auth/password-reset:
    post:
      operationId: resetPassword
      summary: Replace the password using a Supabase recovery access token
      tags: [Identity]
      security: []
      parameters:
        - $ref: '#/components/parameters/CsrfHeader'
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/PasswordResetRequest'
      responses:
        '204':
          description: The password was updated.
        '400':
          $ref: '#/components/responses/ValidationError'
        '401':
          $ref: '#/components/responses/AuthenticationError'
        '422':
          $ref: '#/components/responses/PasswordReuseError'
        '429':
          $ref: '#/components/responses/RateLimitError'
...
  /me:
    get:
      operationId: getMyAccount
      summary: Read the current account and daily quota
      tags: [Identity]
      responses:
        '200':
          description: Current account details.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/AccountResponse'
        '401':
          $ref: '#/components/responses/AuthenticationError'
    delete:
      operationId: deleteMyAccount
      summary: Permanently delete the current account and its activity
      tags: [Identity]
      parameters:
        - $ref: '#/components/parameters/CsrfHeader'
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/DeleteAccountRequest'
      responses:
        '204':
          description: The account, recommendations, votes, and sessions were deleted.
        '400':
          $ref: '#/components/responses/ValidationError'
        '401':
          $ref: '#/components/responses/AuthenticationError'
        '503':
          $ref: '#/components/responses/ServiceUnavailableError'
...
  /admin/reports/users:
    get:
      operationId: getFlaggedUsers
      summary: Search users waiting for report review
      tags: [Moderation]
      parameters:
        - name: query
          in: query
          required: false
          schema:
            type: string
            maxLength: 100
      responses:
        '200':
          description: Flagged users ordered by pending report count.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/FlaggedUsersResponse'
        '401':
          $ref: '#/components/responses/AuthenticationError'
        '404':
          description: Returned when the caller is not an administrator.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ApiErrorResponse'
  /admin/reports/users/{userId}:
    parameters:
      - name: userId
        in: path
        required: true
        schema:
          type: string
          format: uuid
    get:
      operationId: getReportedUser
      summary: Read a user's report history
      tags: [Moderation]
      responses:
        '200':
          description: Reported user and report history.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/UserReportsResponse'
        '401':
          $ref: '#/components/responses/AuthenticationError'
        '404':
          description: The caller is not an administrator or the user does not exist.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ApiErrorResponse'
    patch:
      operationId: moderateReportedUser
      summary: Dismiss pending reports or ban a flagged user
      tags: [Moderation]
      parameters:
        - $ref: '#/components/parameters/CsrfHeader'
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/ModerationRequest'
      responses:
        '200':
          description: Moderation action was applied.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ModerationResult'
        '400':
          $ref: '#/components/responses/ValidationError'
        '401':
          $ref: '#/components/responses/AuthenticationError'
        '403':
          $ref: '#/components/responses/ForbiddenError'
        '404':
          description: The caller is not an administrator or the user does not exist.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ApiErrorResponse'
        '409':
          $ref: '#/components/responses/ConflictError'
...
components:
  securitySchemes:
    sessionCookie:
      type: apiKey
      in: cookie
      name: TRACKDROP_SESSION
  parameters:
    CsrfHeader:
      name: X-XSRF-TOKEN
      in: header
      required: true
      schema:
        type: string
  responses:
    ValidationError:
      description: The request did not satisfy the input policy.
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/ApiErrorResponse'
    ConflictError:
      description: The login ID or normalized email already exists.
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/ApiErrorResponse'
    AuthenticationError:
      description: Authentication is required or the credentials are invalid.
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/ApiErrorResponse'
    PasswordReuseError:
      description: The new password matches the current password.
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/ApiErrorResponse'
    ForbiddenError:
      description: The account or request is not permitted.
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/ApiErrorResponse'
...
    RateLimitError:
      description: The client or server exceeded a protected request limit.
      headers:
        Retry-After:
          schema:
            type: integer
            minimum: 1
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/ApiErrorResponse'
...
    ServiceUnavailableError:
      description: A required external service is temporarily unavailable.
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/ApiErrorResponse'
  schemas:
...
    SignUpResponse:
      type: object
      additionalProperties: false
      required: [emailVerificationRequired]
      properties:
        emailVerificationRequired:
          type: boolean
          const: true
    SignUpRequest:
      type: object
      additionalProperties: false
      required: [email, password]
      properties:
        email:
          type: string
          format: email
          maxLength: 320
        password:
          type: string
          minLength: 8
          maxLength: 16
          format: password
    LoginRequest:
      type: object
      additionalProperties: false
      required: [email, password, rememberMe]
      properties:
        email:
          type: string
          format: email
        password:
          type: string
          format: password
        rememberMe:
          type: boolean
    DeleteAccountRequest:
      type: object
      additionalProperties: false
      required: [password]
      properties:
        password:
          type: string
          format: password
    PasswordRecoveryRequest:
      type: object
      additionalProperties: false
      required: [email]
      properties:
        email:
          type: string
          format: email
          maxLength: 320
    PasswordResetRequest:
      type: object
      additionalProperties: false
      required: [accessToken, password]
      properties:
        accessToken:
          type: string
        password:
          type: string
          minLength: 8
          maxLength: 16
          format: password
    AccountResponse:
      type: object
      additionalProperties: false
      required: [account, quota]
      properties:
        account:
          $ref: '#/components/schemas/Account'
        quota:
          $ref: '#/components/schemas/DailyQuota'
    Account:
      type: object
      additionalProperties: false
      required: [email, publicNickname, emailVerified, createdAt, admin]
      properties:
        email:
          type: string
          format: email
        publicNickname:
          type: string
        emailVerified:
          type: boolean
        createdAt:
          type: string
          format: date-time
        admin:
          type: boolean
    DailyQuota:
      type: object
      additionalProperties: false
      required: [date, limit, used, remaining, resetAt]
      properties:
        date:
          type: string
          format: date
        limit:
          type: integer
          const: 4
        used:
          type: integer
          minimum: 0
          maximum: 4
        remaining:
          type: integer
          minimum: 0
          maximum: 4
        resetAt:
          type: string
          format: date-time
...
    FlaggedUsersResponse:
      type: object
      additionalProperties: false
      required: [items]
      properties:
        items:
          type: array
          maxItems: 50
          items:
            $ref: '#/components/schemas/FlaggedUser'
    FlaggedUser:
      type: object
      additionalProperties: false
      required: [id, email, publicNickname, activity, pendingReportCount, firstReportedAt, latestReportedAt]
      properties:
        id:
          type: string
          format: uuid
        email:
          type: string
          format: email
        publicNickname:
          type: string
        activity:
          type: string
          const: FLAGGED
        pendingReportCount:
          type: integer
          minimum: 1
        firstReportedAt:
          type: string
          format: date-time
        latestReportedAt:
          type: string
          format: date-time
    UserReportsResponse:
      type: object
      additionalProperties: false
      required: [user, reports]
      properties:
        user:
          $ref: '#/components/schemas/ReportedUser'
        reports:
          type: array
          items:
            $ref: '#/components/schemas/AdminReportRecord'
    ReportedUser:
      type: object
      additionalProperties: false
      required: [id, email, publicNickname, activity, pendingReportCount]
      properties:
        id:
          type: string
          format: uuid
        email:
          type: string
          format: email
        publicNickname:
          type: string
        activity:
          type: string
          enum: [NORMAL, FLAGGED, BAN]
        pendingReportCount:
          type: integer
          minimum: 0
    AdminReportRecord:
      type: object
      additionalProperties: false
      required: [id, reasonCode, details, status, createdAt, reporter, recommendation]
      properties:
        id:
          type: string
          format: uuid
        reasonCode:
          type: string
          enum: [ABUSIVE_LANGUAGE, SPAM, OTHER]
        details:
          type: [string, 'null']
        status:
          type: string
          enum: [PENDING, REVIEWED, DISMISSED, ACTIONED]
        createdAt:
          type: string
          format: date-time
        reporter:
          $ref: '#/components/schemas/AdminReporter'
        recommendation:
          $ref: '#/components/schemas/AdminRecommendation'
    AdminReporter:
      type: object
      additionalProperties: false
      required: [id, publicNickname]
      properties:
        id:
          type: string
          format: uuid
        publicNickname:
          type: string
    AdminRecommendation:
      type: object
      additionalProperties: false
      required: [id, comment, trackId, trackTitle, artistName]
      properties:
        id:
          type: string
          format: uuid
        comment:
          type: string
        trackId:
          type: string
          format: uuid
        trackTitle:
          type: string
        artistName:
          type: string
    ModerationRequest:
      type: object
      additionalProperties: false
      required: [action]
      properties:
        action:
          type: string
          enum: [DISMISS, BAN]
    ModerationResult:
      type: object
      additionalProperties: false
      required: [userId, activity, accountStatus, resolvedReportCount, resolvedAt]
      properties:
        userId:
          type: string
          format: uuid
        activity:
          type: string
          enum: [NORMAL, BAN]
        accountStatus:
          type: string
          enum: [ACTIVE, SUSPENDED, WITHDRAWN]
        resolvedReportCount:
          type: integer
          minimum: 0
        resolvedAt:
          type: string
          format: date-time
...
    ApiErrorResponse:
      type: object
      additionalProperties: false
      required: [error]
      properties:
        error:
          $ref: '#/components/schemas/ApiError'
    ApiError:
      type: object
      additionalProperties: false
      required: [code, message]
      properties:
        code:
          type: string
        message:
          type: string
        details:
          type: object
          additionalProperties: true
security:
  - sessionCookie: []